gaia improvements
