+++ src/inspect_evals/gaia/__init__.py
@@ -1,15 +1,17 @@
+from .dataset import gaia_dataset
 from .gaia import (
     gaia,
-    gaia_dataset,
     gaia_level1,
     gaia_level2,
     gaia_level3,
 )
+from .scorer import gaia_scorer
 
 __all__ = [
     "gaia",
     "gaia_level1",
     "gaia_level2",
     "gaia_level3",
+    "gaia_scorer",
     "gaia_dataset",
 ]
